fix mypy

+++ app/models/sleep.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Integer, ForeignKey
+from sqlalchemy import Float, Integer, ForeignKey
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 from .base import Base
 
@@ -10,6 +10,6 @@ class Sleep(Base):
     user_telegram_id: Mapped[int] = mapped_column(
         ForeignKey("users.telegram_id", ondelete="CASCADE"), unique=True, nullable=False
     )
-    hours: Mapped[int] = mapped_column(Integer, default=0)
+    hours: Mapped[float] = mapped_column(Float, default=0)
 
     user: Mapped["User"] = relationship(back_populates="sleep")  # type: ignore  # noqa: F821
