+++ app/schemas/sleep.py
@@ -3,7 +3,7 @@ from pydantic import BaseModel
 
 class SleepSchema(BaseModel):
     user_telegram_id: int
-    hours: int
+    hours: float
 
     class Config:
         orm_mode = True
